pyproject.toml
[tool.poetry]
name = "qachat"
version = "0.1.0"
description = ""
authors = ["wajidminhas <shanitent667@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.36.0"
google-generativeai = "^0.7.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
